+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - A warning is printed when `CALL` and `RETURN` are used with atypical registers.
 - `help` debugging command takes one or more arguments and prints more detailed help messages for specific commands.
 - `TIGER_STRING` is accepted as an alias for `LP_STRING`.
+- `--big-stack` command-line flag.
 
 ### Changed
 - Warning for zero-prefixed octal literals will only be printed once per program, rather than for every occurrence.